+++ src/index.ts
@@ -1,1 +1,5 @@
-export { Express, RequestExtended, ExpressCallback, Express404Callback, ExpressListeningCallback, SuccessResponse, FailureResponse } from "./express.service"
+import express from "express"
+export { CorsOptions } from "cors"
+export { HelmetOptions } from "helmet"
+export { express }
+export { Express, RequestExtended, Express404InterceptorCallback, ExpressErrorInterceptorCallback, ExpressFailedResponseCallback, ExpressResponseCallback, ExpressListeningCallback, SuccessResponse, FailureResponse } from "./express.service"